+++ pyincore/dfr3service.py
@@ -5,13 +5,13 @@
 # and is available at https://www.mozilla.org/en-US/MPL/2.0/
 
 
+import re
 import urllib
 from typing import Dict
 
-import jsonpickle
-import re
-
 from pyincore import IncoreClient
+from pyincore.models.fragilitycurveset import FragilityCurveSet
+from pyincore.mappingset import MappingSet
 
 
 class MappingSubject(object):
@@ -48,67 +48,6 @@ class Dfr3Service:
         self.base_mapping_url = urllib.parse.urljoin(client.service_url,
                                                      'dfr3/api/mappings/')
 
-    def map_inventory(self, mapping_id: str, inventories: dict, key: str):
-        """Mapping between inventories (buildings, bridges etc.) and DFR3 sets.
-
-        Args:
-            mapping_id (str): ID of the Mapping file.
-            inventories (dict):  Infrastructure inventory.
-            key (str): Parameter's key in param dictionary.
-
-        Returns:
-            dict: DFR3 sets from the response.
-
-        """
-        features = []
-
-        for inventory in inventories:
-            # hack, change null to an empty string
-            # Some metadata field values are null even though the field is defined
-            # in the Mapping as a string
-            if "occ_type" in inventory["properties"] and \
-                    inventory["properties"]["occ_type"] is None:
-                inventory["properties"]["occ_type"] = ""
-            if "efacility" in inventory["properties"] and \
-                    inventory["properties"]["efacility"] is None:
-                inventory["properties"]["efacility"] = ""
-
-            features.append(inventory)
-
-        feature_collection = {
-            "type": "FeatureCollection",
-            "features": features,
-        }
-
-        mapping_request = MappingRequest()
-        mapping_request.subject.schema = "building"  # currently it is not used in the service side
-        mapping_request.subject.inventory = feature_collection
-        mapping_request.params["key"] = key
-
-        url = urllib.parse.urljoin(self.base_mapping_url,
-                                   mapping_id + "/matched")
-
-        json = jsonpickle.encode(mapping_request, unpicklable=False).encode(
-            "utf-8")
-        headers = {'Content-type': 'application/json'}
-        # merge two headers
-        new_headers = {**self.client.session.headers, **headers}
-        kwargs = {"headers": new_headers}
-        r = self.client.post(url, data=json, **kwargs)
-
-        response = r.json()
-
-        # construct list of DFR3 sets
-        mapping = response["mapping"]
-        sets = response["sets"]
-
-        # reconstruct a dictionary of DFR3 sets from the response
-        dfr3_sets = {}
-        for k, v in mapping.items():
-            dfr3_sets[k] = sets[v]
-
-        return dfr3_sets
-
     def get_dfr3_set(self, dfr3_id: str):
         """Get all DFR3 sets.
 
@@ -128,6 +67,22 @@ class Dfr3Service:
         batch_dfr3_sets = {}
         for id in dfr3_id_lists:
             batch_dfr3_sets[id] = self.get_dfr3_set(id)
+
+        return batch_dfr3_sets
+
+    def batch_get_dfr3_set_object(self, dfr3_id_lists: list):
+        """
+        this method is intended to replace batch_get_dfr3_set in the future
+        It retrieve dfr3 sets from services using id and instantiate DFR3Curveset objects in bulk
+        Args:
+            dfr3_id_lists: list of ids
+
+        Returns: list of dfr3curve object
+
+        """
+        batch_dfr3_sets = {}
+        for id in dfr3_id_lists:
+            batch_dfr3_sets[id] = FragilityCurveSet(self.get_dfr3_set(id))
 
         return batch_dfr3_sets
 
@@ -202,6 +157,57 @@ class Dfr3Service:
 
         return dfr3_sets
 
+    def match_inventory_object(self, mapping:MappingSet, inventories: dict, entry_key: str):
+        """
+        This method is intended to replace the match_inventory method in the future. The functionality is same as
+        match_inventory but instead of dfr3_sets in plain json, dfr3 curves will be represented in
+        FragilityCurveSet Object
+        Args:
+            mapping: MappingSet Object that has the rules and entries
+            inventories: inventories
+            entry_key: keys such as PGA, pgd, and etc
+
+        Returns: a dictionary of {"inventory id": FragilityCurveSet object}
+
+        """
+        dfr3_sets = {}
+
+        # 2. loop through inventory to match the rules
+        matched_curve_ids = []
+        for inventory in inventories:
+            if "occ_type" in inventory["properties"] and \
+                    inventory["properties"]["occ_type"] is None:
+                inventory["properties"]["occ_type"] = ""
+            if "efacility" in inventory["properties"] and \
+                    inventory["properties"]["efacility"] is None:
+                inventory["properties"]["efacility"] = ""
+
+            for m in mapping.mappings:
+
+                if self._property_match(rules=m.rules, properties=inventory["properties"]):
+                    curve = m.entry[entry_key]
+                    dfr3_sets[inventory['id']] = curve
+
+                    # if it's string:id; then need to fetch it from remote and cast to fragility3curve object
+                    if isinstance(curve, str) and curve not in matched_curve_ids:
+                        matched_curve_ids.append(curve)
+
+                    # use the first match
+                    break
+
+        batch_dfr3_sets = self.batch_get_dfr3_set_object(matched_curve_ids)
+
+        # 3. replace the curve id in dfr3_sets to the dfr3 curve
+        for inventory_id, curve_item in dfr3_sets.items():
+            if isinstance(curve_item, FragilityCurveSet):
+                pass
+            elif isinstance(curve_item, str):
+                dfr3_sets[inventory_id] = batch_dfr3_sets[curve_item]
+            else:
+                raise ValueError("Cannot realize dfr3_set entry. The entry has to be either remote id string; or dfr3curve object!")
+
+        return dfr3_sets
+
     def _property_match(self, rules, properties):
 
         # add more types if needed
@@ -225,8 +231,8 @@ class Dfr3Service:
             "MATCHES": ""
         }
 
-        # if rules is [[]] meaning it matches without any condition
-        if rules == [[]]:
+        # if rules match without any condition
+        if rules == [[]] or rules == [] or rules == [None]:
             return True
 
         else:
@@ -251,19 +257,23 @@ class Dfr3Service:
 
                 rule_value = elements[3].strip('\'').strip('\"')
 
-                if rule_key in properties.keys() and isinstance(properties[rule_key],
-                                                                eval(known_types[rule_type])):
-                    if rule_type == 'java.lang.String':
-                        if rule_operator == "MATCHES":
-                            matched = bool(re.search(rule_value, properties[rule_key]))
-                        elif rule_operator == "NMATCHES":
-                            matched = not bool(re.search(rule_value, properties[rule_key]))
+                if rule_key in properties.keys():
+                    if isinstance(properties[rule_key], eval(known_types[rule_type])):
+                        if rule_type == 'java.lang.String':
+                            if rule_operator == "MATCHES":
+                                matched = bool(re.search(rule_value, properties[rule_key]))
+                            elif rule_operator == "NMATCHES":
+                                matched = not bool(re.search(rule_value, properties[rule_key]))
+                            else:
+                                matched = eval(
+                                    '"{0}"'.format(properties[rule_key]) + known_operators[rule_operator] + '"{0}"'.format(
+                                        rule_value))
                         else:
-                            matched = eval(
-                                '"{0}"'.format(properties[rule_key]) + known_operators[rule_operator] + '"{0}"'.format(
-                                    rule_value))
+                            matched = eval(str(properties[rule_key]) + known_operators[rule_operator] + rule_value)
                     else:
-                        matched = eval(str(properties[rule_key]) + known_operators[rule_operator] + rule_value)
+                        raise ValueError("Mismatched datatype found in the mapping rule: " + rule +
+                                         ". Datatype found in the dataset for " + rule_key + " : "
+                                         + str(type(properties[rule_key])) + ". Please review the mapping being used.")
 
                 if not matched:
                     break