+++ pyincore/utils/analysisutil.py
@@ -110,7 +110,7 @@ class AnalysisUtil:
                         hazard, period, fs_param0,
                         fs_param1, fs_param2, fs_param3,
                         fs_param4, fs_param5)
-                elif class_name in ['PeriodStandardFragilityCurve','StandardFragilityCurve']:
+                elif class_name in ['PeriodStandardFragilityCurve', 'StandardFragilityCurve']:
                     alpha = float(fragility_curve['alpha'])
                     beta = math.sqrt(math.pow(fragility_curve["beta"], 2) + math.pow(std_dev, 2))
 
@@ -548,6 +548,57 @@ class AnalysisUtil:
         return hazard_demand_type
 
     @staticmethod
+    def get_hazard_demand_type_object(building, fragility_set, hazard_type):
+        """
+        Get hazard demand type. This method is intended to replace get_hazard_demand_type. Fragility_set is not a
+        json but a fragilityCurveSet object now.
+
+        Args:
+            building (obj): A JSON mapping of a geometric object from the inventory: current building.
+            fragility_set (obj): FragilityCurveSet object
+            hazard_type (str): A hazard type such as earthquake, tsunami etc.
+
+        Returns:
+            str: A hazard demand type.
+
+        """
+        BLDG_STORIES = "no_stories"
+        PROPERTIES = "properties"
+        BLDG_PERIOD = "period"
+
+        fragility_hazard_type = fragility_set.demand_type.lower()
+        hazard_demand_type = fragility_hazard_type
+
+        if hazard_type.lower() == "earthquake":
+            num_stories = building[PROPERTIES][BLDG_STORIES]
+            # Get building period from the fragility if possible
+            building_period = fragility_set.fragility_curves[0].get_building_period(num_stories)
+
+            if fragility_hazard_type.endswith('sa') and fragility_hazard_type != 'sa':
+                # This fixes a bug where demand type is in a format similar to 1.0 Sec Sa
+                if len(fragility_hazard_type.split()) > 2:
+                    building_period = fragility_hazard_type.split()[0]
+                    fragility_hazard_type = "Sa"
+
+            hazard_demand_type = fragility_hazard_type
+
+            # This handles the case where some fragilities only specify Sa, others a specific period of Sa
+            if not hazard_demand_type.endswith('pga'):
+                # If the fragility does not contain the period calculation, check if the dataset has it
+                if building_period == 0.0 and BLDG_PERIOD in building[PROPERTIES]:
+                    if building[PROPERTIES][BLDG_PERIOD] > 0.0:
+                        building_period = building[PROPERTIES][BLDG_PERIOD]
+
+                hazard_demand_type = str(building_period) + " " + fragility_hazard_type
+        elif hazard_type.lower() == "tsunami":
+            if hazard_demand_type == "momentumflux":
+                hazard_demand_type = "mmax"
+            elif hazard_demand_type == "inundationdepth":
+                hazard_demand_type = "hmax"
+
+        return hazard_demand_type
+
+    @staticmethod
     def group_by_demand_type(inventories, fragility_sets, hazard_type="earthquake", is_building=False):
         """
 
@@ -569,6 +620,35 @@ class AnalysisUtil:
             if is_building:
                 inventory = inventories[fragility_id]
                 demand_type = AnalysisUtil.get_hazard_demand_type(inventory, frag, hazard_type)
+
+            tpl = (demand_type, demand_units)
+            grouped_inventory.setdefault(tpl, []).append(fragility_id)
+
+        return grouped_inventory
+
+    @staticmethod
+    def group_by_demand_type_object(inventories, fragility_sets, hazard_type="earthquake", is_building=False):
+        """
+        This method should replace group_by_demand_type in the future. Fragility_sets is not list of dictionary (
+        json) anymore but a list of FragilityCurveSet objects
+        Args:
+            inventories: dictionary of {id: intentory}
+            fragility_sets: fragility_sets
+            hazard_type: default to earthquake
+            is_building: if the inventory is builiding or not
+
+        Returns: dictionary of grouped inventory with { (demandunit, demandtype):[inventory ids] }
+
+        """
+        grouped_inventory = dict()
+        for fragility_id, frag in fragility_sets.items():
+
+            demand_type = frag.demand_type
+            demand_units = frag.demand_units
+
+            if is_building:
+                inventory = inventories[fragility_id]
+                demand_type = AnalysisUtil.get_hazard_demand_type_object(inventory, frag, hazard_type)
 
             tpl = (demand_type, demand_units)
             grouped_inventory.setdefault(tpl, []).append(fragility_id)